refactor: 커스텀컬러, 에셋 경로 정리
